[fix]1つになっていたコンポーネントを2つに分けた
