+++ services/nutfesmap-web/src/components/map/PlanPin.tsx
@@ -1,24 +1,14 @@
 "use client";
 
-import React, { useState } from "react";
-import { FaHamburger } from "react-icons/fa";
-import { FaMapPin, FaBuildingColumns } from "react-icons/fa6";
+import React from "react";
+import { FaMapPin } from "react-icons/fa6";
 import { MdFamilyRestroom } from "react-icons/md";
-import BottomSheet from "@/components/plans/BottomSheet";
+import { FaHamburger } from "react-icons/fa";
+import { FaBuildingColumns } from "react-icons/fa6";
 import { Category } from "@/types/enums";
 
-// --- SpotData型（BottomSheetに渡す用） ---
-type SpotData = {
-  title: string;
-  category: Category;
-  time: string;
-  location: string;
-  description: string;
-  imageUrl: string;
-};
-
 // --- Pin型 ---
-type Pin = {
+export type Pin = {
   id: number;
   top: string;
   left: string;
@@ -27,34 +17,6 @@ type Pin = {
   waitTime: number;
 };
 
-// --- ピンのデータ ---
-const pinData: Pin[] = [
-  {
-    id: 1,
-    top: "50px",
-    left: "150px",
-    category: Category.Food,
-    eventName: "スイートポテトコンテスト",
-    waitTime: 15,
-  },
-  {
-    id: 2,
-    top: "250px",
-    left: "80px",
-    category: Category.Child,
-    eventName: "こども広場",
-    waitTime: 30,
-  },
-  {
-    id: 3,
-    top: "200px",
-    left: "280px",
-    category: Category.Plan,
-    eventName: "特別企画展",
-    waitTime: 5,
-  },
-];
-
 // --- カテゴリアイコン ---
 const categoryIcons: Record<Category, React.ComponentType> = {
   [Category.Food]: FaHamburger,
@@ -62,11 +24,12 @@ const categoryIcons: Record<Category, React.ComponentType> = {
   [Category.Plan]: FaBuildingColumns,
 };
 
-// --- 個々のピン ---
-const PlanPin: React.FC<{ pin: Pin; onClick: (pin: Pin) => void }> = ({
-  pin,
-  onClick,
-}) => {
+type PlanPinProps = {
+  pin: Pin;
+  onClick: (pin: Pin) => void;
+};
+
+const PlanPin: React.FC<PlanPinProps> = ({ pin, onClick }) => {
   const CategoryIcon = categoryIcons[pin.category];
 
   return (
@@ -93,41 +56,4 @@ const PlanPin: React.FC<{ pin: Pin; onClick: (pin: Pin) => void }> = ({
   );
 };
 
-// --- 全ピン表示 + BottomSheet制御 ---
-const PlanPinsOnMap = () => {
-  const [selectedPin, setSelectedPin] = useState<Pin | null>(null);
-  const [isOpen, setIsOpen] = useState(false);
-
-  const handlePinClick = (pin: Pin) => {
-    setSelectedPin(pin);
-    setIsOpen(true);
-  };
-
-  const handleClose = () => setIsOpen(false);
-
-  // ピン情報 → BottomSheet 用データに変換
-  const spotData: SpotData | null = selectedPin
-    ? {
-        title: selectedPin.eventName,
-        category: selectedPin.category,
-        time: `${selectedPin.waitTime}分`,
-        location: "未設定",
-        description: "このイベントの詳細情報です。",
-        imageUrl: "/example.png",
-      }
-    : null;
-
-  return (
-    <div className="relative w-full h-screen bg-gray-100">
-      {pinData.map((pin) => (
-        <PlanPin key={pin.id} pin={pin} onClick={handlePinClick} />
-      ))}
-
-      {spotData && (
-        <BottomSheet isOpen={isOpen} onClose={handleClose} spotData={spotData} />
-      )}
-    </div>
-  );
-};
-
-export default PlanPinsOnMap;
+export default PlanPin;
